# Input dimension checking when evaluating surrogates