--- a/lib/SurrogatesPolyChaos/src/SurrogatesPolyChaos.jl
+++ b/lib/SurrogatesPolyChaos/src/SurrogatesPolyChaos.jl
@@ -1,6 +1,6 @@
 module SurrogatesPolyChaos
 
-import Surrogates: AbstractSurrogate, add_point!
+import Surrogates: AbstractSurrogate, add_point!, _check_dimension
 export PolynomialChaosSurrogate
 
 using PolyChaos
@@ -37,6 +37,9 @@ function PolynomialChaosSurrogate(x, y, lb::Number, ub::Number;
 end
 
 function (pc::PolynomialChaosSurrogate)(val::Number)
+    # Check to make sure dimensions of input matches expected dimension of surrogate
+    _check_dimension(pc, val)
+
     return sum([pc.coeff[i] * PolyChaos.evaluate(val, pc.ortopolys)[i]
                 for i in 1:(pc.num_of_multi_indexes)])
 end
@@ -71,6 +74,9 @@ function PolynomialChaosSurrogate(x, y, lb, ub;
 end
 
 function (pcND::PolynomialChaosSurrogate)(val)
+    # Check to make sure dimensions of input matches expected dimension of surrogate
+    _check_dimension(pcND, val)
+
     sum = zero(eltype(val[1]))
     for i in 1:(pcND.num_of_multi_indexes)
         sum = sum +