--- a/src/LinearSurrogate.jl
+++ b/src/LinearSurrogate.jl
@@ -51,6 +51,8 @@ function add_point!(my_linear::LinearSurrogate, new_x, new_y)
 end
 
 function (lin::LinearSurrogate)(val)
+    # Check to make sure dimensions of input matches expected dimension of surrogate
+    _check_dimension(lin, val)
     return lin.coeff' * [val...]
 end