--- a/src/utils.jl
+++ b/src/utils.jl
@@ -3,8 +3,10 @@ remove_tracker(x) = x
 _match_container(y, y_el::Number) = first(y)
 _match_container(y, y_el) = y
 
+_expected_dimension(x) = length(x[1])
+
 function _check_dimension(surr, input)
-    expected_dim = length(surr.x[1])
+    expected_dim = _expected_dimension(surr.x)
     input_dim = length(input)
 
     if input_dim != expected_dim
@@ -12,6 +14,5 @@ function _check_dimension(surr, input)
             ArgumentError("This surrogate expects $expected_dim-dimensional inputs, but the input had dimension $input_dim.")
         )
     end
-
     return nothing
 end