--- a/test/GEK.jl
+++ b/test/GEK.jl
@@ -16,6 +16,11 @@ val = my_gek(2.0)
 std_err = std_error_at_point(my_gek, 1.0)
 add_point!(my_gek, 2.5, 2.5^2)
 
+# Test that input dimension is properly checked for 1D GEK surrogates
+@test_throws ArgumentError my_gek(Float64[])
+@test_throws ArgumentError my_gek((2.0, 3.0, 4.0))
+
+
 #ND
 n = 10
 d = 2
@@ -43,3 +48,8 @@ my_gek_ND = GEK(x, y, lb, ub)
 val = my_gek_ND((1.0, 1.0))
 std_err = std_error_at_point(my_gek_ND, (1.0, 1.0))
 add_point!(my_gek_ND, (2.0, 2.0), 8.0)
+
+# Test that input dimension is properly checked for ND GEK surrogates
+@test_throws ArgumentError my_gek_ND(Float64[])
+@test_throws ArgumentError my_gek_ND(2.0)
+@test_throws ArgumentError my_gek_ND((2.0, 3.0, 4.0))