… test/Wendland.jl
… test/earth.jl
… test/inverseDistanceSurrogate.jl
… test/linearSurrogate.jl
… test/lobachevsky.jl
… test/secondOrderPolynomialSurrogate.jl