Fixes https://github.com/SciML/Surrogates.jl/issues/337. Adds a new utility function `_check_input_dimension` which is called for all surrogates before evaluating a point. Provides a helpful error message if there's a dimension mismatch. 

# Example

```julia
using Surrogates

lb = 0.0
ub = 4.0
x = [1.0,2.0,3.0]
y = [4.0,5.0,6.0]
my_rad = RadialBasis(x, y, lb, ub, rad = linearRadial())
```

```julia
julia> my_rad(Float64[1.5])
4.5

julia> my_rad([1.0, 2.0])
ERROR: ArgumentError: This surrogate expects 1-dimensional inputs, but the input had dimension 2.
Stacktrace:
 [1] _check_dimension
   @ C:\Users\thoma\Surrogates.jl\src\utils.jl:11 [inlined]
 [2] (::RadialBasis{Surrogates.var"#1#2", Int64, Vector{Float64}, Vector{Float64}, Float64, Float64, Transpose{Float64, Vector{Float64}}, Float64, Bool})(val::Vector{Float64})
   @ Surrogates C:\Users\thoma\Surrogates.jl\src\Radials.jl:159
 [3] top-level scope
   @ REPL[39]:1

julia> my_rad(Float64[])
ERROR: ArgumentError: This surrogate expects 1-dimensional inputs, but the input had dimension 0.
Stacktrace:
 [1] _check_dimension
   @ C:\Users\thoma\Surrogates.jl\src\utils.jl:11 [inlined]
 [2] (::RadialBasis{Surrogates.var"#1#2", Int64, Vector{Float64}, Vector{Float64}, Float64, Float64, Transpose{Float64, Vector{Float64}}, Float64, Bool})(val::Vector{Float64})
   @ Surrogates C:\Users\thoma\Surrogates.jl\src\Radials.jl:159
 [3] top-level scope
   @ REPL[40]:1
```

Benchmarking reveals this to come at a very small cost:

**Before**
julia> @benchmark my_rad([1.5])
BenchmarkTools.Trial: 10000 samples with 855 evaluations.
 Range (min … max):  143.392 ns …  10.876 μs  ┊ GC (min … max): 0.00% … 97.94%
 Time  (median):     146.784 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   175.408 ns ± 332.221 ns  ┊ GC (mean ± σ):  7.69% ±  4.01%

  ▆█▆▅▄▂▁                                   ▁  ▂▄▄▂    ▁        ▁
  ████████▆▆▇▅▅▅▆▅▅▅▄▅▆▅▄▄▃▅▅▅▅▅▅▅▅▄▄▅▅▅▅▄▇█████████▇▆██▇▆▅▄▅▄▄ █
  143 ns        Histogram: log(frequency) by time        247 ns <

 Memory estimate: 144 bytes, allocs estimate: 3.

**After**
julia> @benchmark my_rad([1.5])
BenchmarkTools.Trial: 10000 samples with 825 evaluations.
 Range (min … max):  151.394 ns …  10.147 μs  ┊ GC (min … max): 0.00% … 97.38%
 Time  (median):     155.394 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   196.150 ns ± 340.000 ns  ┊ GC (mean ± σ):  6.83% ±  3.89%

  ▇█▄▃▂▂                       ▁▁▁  ▂▄▅▅▂▁▁▂                    ▁
  ████████▇▆▅▆▇▇▇▆▆▇▆▇▆▆▆▆▆▅▅▅▇██████████████▆▄▅▅▄▅▆▅▅▆▆▅▆▄▃▅▄▄ █
  151 ns        Histogram: log(frequency) by time        302 ns <

 Memory estimate: 144 bytes, allocs estimate: 3.