@archermarx 2022-07-13
I have removed the tests for GEKPLS for now because it expects matrix inputs instead of vectors of tuples like the rest of Surrogates.jl. When https://github.com/SciML/Surrogates.jl/issues/355 is being worked on, I will revisit it